+++ pandapower/test/api/test_group.py
@@ -10,6 +10,7 @@ import pandas as pd
 import pandas.testing as pdt
 from packaging.version import Version
 
+import pandapower.toolbox
 from pandapower._version import __version__
 import pandapower as pp
 import pandapower.networks as nw
@@ -31,7 +32,7 @@ def nets_to_test_group():
     idxs = list()
     for reference_column, type_ in zip(reference_columns, types):
         net = nw.case24_ieee_rts()
-        for elm in pp.pp_elements():
+        for elm in pandapower.toolbox.pp_elements():
             net[elm]["name"] = np.arange(net[elm].shape[0]).astype(str)
         idx0 = pp.create_group_from_dict(net, {
             "gen": typed_list([0, 1], type_),
@@ -77,9 +78,9 @@ def test_group_element_index():
     for net, type_, rc, idxs in zip(*nets_to_test_group()):
 
         # ! group_element_index()
-        assert (pp.group_element_index(net, 0, "gen") == pd.Index([0, 1], dtype=int)).all()
-        assert (pp.group_element_index(net, 0, "sgen") == pd.Index([2, 3], dtype=int)).all()
-        assert (pp.group_element_index(net, 0, "dcline") == pd.Index([], dtype=int)).all()
+        assert (pp.group_element_index(net, 0, "gen") == pd.Index([0, 1], dtype=np.int64)).all()
+        assert (pp.group_element_index(net, 0, "sgen") == pd.Index([2, 3], dtype=np.int64)).all()
+        assert (pp.group_element_index(net, 0, "dcline") == pd.Index([], dtype=np.int64)).all()
 
 
 def test_groups_equal():
@@ -179,6 +180,45 @@ def test_remove_not_existing_group_members():
         assert "gen" not in net.group.element_type.loc[[idxs[1]]].values
 
 
+def test_check_unique_group_rows():
+    net = pp.create_empty_network()
+    empty_group = deepcopy(net.group)
+
+    # test with duplicated rows
+    net.group = pd.concat([empty_group, pd.DataFrame([
+        ["Gr1",  "gen", [1, 2]],
+        ["Gr1", "sgen", [3, 4]],
+        ["Gr1",  "gen", [2, 5]],
+    ], index=[0, 0, 0], columns=["name", "element_type", "element"])])
+    try:
+        pp.check_unique_group_rows(net)
+        assert False, "ValueError expected"
+    except ValueError:
+        pass
+
+    # test with different reference_columns
+    net.group.iat[0, 3] = "hallo"
+    try:
+        pp.check_unique_group_rows(net)
+        assert False, "ValueError expected"
+    except ValueError:
+        pass
+
+    # test with duplicated group name and index
+    net.group = pd.concat([empty_group, pd.DataFrame([
+        ["Gr1",  "gen", [1, 2]],
+        ["Gr1", "sgen", [3, 4]],
+        ["Gr2",  "gen", [2, 5]],
+        ["Gr3", "line", [0, 1]]
+    ], index=[0, 0, 1, 0], columns=["name", "element_type", "element"])])
+    try:
+        pp.check_unique_group_rows(net)
+        assert False, "UserWarning expected"
+    except UserWarning:
+        pass
+    pp.check_unique_group_rows(net, raise_=False, log_level="debug")
+
+
 def test_drop_element():
     net = nw.case24_ieee_rts()
     gr1 = pp.create_group_from_dict(net, {
@@ -229,7 +269,7 @@ def test_drop_and_return():
             assert net2b.gen.shape[0] == 10  # unchanged
             assert net2b.trafo.shape[0] == 2
             assert set(net3.gen.index) == {0, 1}
-            for elm in pp.pp_elements():
+            for elm in pandapower.toolbox.pp_elements():
                 assert net2[elm].shape[0] <= net[elm].shape[0]
                 assert net2b[elm].shape[0] <= net[elm].shape[0]
                 assert set(net2[elm].index) | set(net3[elm].index) == set(net[elm].index)
@@ -305,6 +345,29 @@ def test_res_power():
             net.res_sgen.p_mw.loc[[2, 3]].sum()
         assert np.isclose(pp.group_res_p_mw(net, 3), p_val)
 
+        # compare per_bus and sum
+        for gr_idx in net.group.index:
+            per_bus_out = pp.group_res_power_per_bus(net, gr_idx)
+            assert np.isclose(per_bus_out.sum().at["p_mw"], pp.group_res_p_mw(net, gr_idx))
+            assert np.isclose(per_bus_out.sum().at["q_mvar"], pp.group_res_q_mvar(net, gr_idx))
+            if gr_idx == 3:
+                assert isinstance(per_bus_out, pd.DataFrame)
+                assert per_bus_out.columns.tolist() == ["p_mw", "q_mvar"]
+                assert per_bus_out.index.tolist() == [0, 1, 2, 8, 10, 11, 23]
+
+
+def test_res_power_examples():
+    net = nw.create_cigre_network_mv(with_der="all")
+    pp.runpp(net)
+    idx = pp.create_group(net, ["sgen", "line"], [[0, 1], [0, 1]], name="test group")
+    expected = pd.DataFrame([
+        [ 2.953004,  1.328978],
+        [ 0.      ,  0.      ],
+        [-2.875066, -1.318864],
+        [-0.02    ,  0.      ]
+        ], index=pd.Index([1, 2, 3, 4], name="bus"), columns=["p_mw", "q_mvar"])
+    assert pandapower.toolbox.dataframes_equal(pp.group_res_power_per_bus(net, idx), expected, atol=1e-6)
+
 
 def test_group_io():
     net = nw.case24_ieee_rts()
@@ -325,10 +388,10 @@ def test_count_group_elements():
     for net, type_, rc, idxs in zip(*nets_to_test_group()):
         pdt.assert_series_equal(
             pp.count_group_elements(net, idxs[0]),
-            pd.Series({"gen": 2, "sgen": 2}, dtype=int))
+            pd.Series({"gen": 2, "sgen": 2}, dtype=np.int64))
         pdt.assert_series_equal(
             pp.count_group_elements(net, idxs[1]),
-            pd.Series({"trafo": 3}, dtype=int))
+            pd.Series({"trafo": 3}, dtype=np.int64))
 
 
 def test_isin():
@@ -338,6 +401,18 @@ def test_isin():
         assert pp.isin_group(net, "gen", 0)
         assert not pp.isin_group(net, "gen", 0, index=idxs[1])
         assert not pp.isin_group(net, "gen", 6)
+
+
+def test_element_associated_groups():
+    for net, type_, rc, idxs in zip(*nets_to_test_group()):
+        assert pp.element_associated_groups(net, "gen", [0, 1, 2, 3]) == \
+            {0: [0], 1: [0], 2: [], 3: []}
+        assert pp.element_associated_groups(net, "gen", [0, 1, 2, 3], return_empties=False) == \
+            pp.element_associated_groups(net, "gen", net.gen.index, return_empties=False) == \
+            {0: [0], 1: [0]}
+        assert pp.element_associated_groups(net, "load", [0, 1]) == {0: [], 1: []}
+        assert pp.element_associated_groups(net, "trafo", [0, 1, 3]) == {0: [3], 1: [3], 3: []}
+        assert pp.element_associated_groups(net, "trafo", 0) == [3]
 
 
 def test_elements_connected_to_group():
@@ -391,24 +466,4 @@ def test_elements_connected_to_group():
 
 
 if __name__ == "__main__":
-    if 0:
-        pytest.main(['-x', "test_group.py"])
-    else:
-        # test_group_create()
-        # test_group_element_index()
-        # test_groups_equal()
-        # test_set_group_reference_column()
-        # test_compare_group_elements()
-        # test_ensure_lists_in_group_element_column()
-        # test_remove_not_existing_group_members()
-        # test_drop_element()
-        # test_drop_and_return()
-        # test_set_out_of_service()
-        # test_attach_to_group()
-        # test_detach_and_compare()
-        # test_res_power()
-        # test_group_io()
-        # test_count_group_elements()
-        # test_isin()
-        test_elements_connected_to_group()
-        pass
+    pytest.main([__file__, "-xs"])